requirements.txt
aiohttp==3.6.2
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
Django==3.0.7
django-filter==2.3.0
djangorestframework==3.11.0
idna==2.10
multidict==4.7.6
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
yarl==1.4.2
